+++ scripts/prepareTests.bash
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-
-yarn build:link \
-  && cd test/project \
-  && yarn \
-  && yarn link serverless-openapi-documentation