+++ src/utils.ts
@@ -3,7 +3,7 @@ import { Clone, IMerge, Merge } from 'lutils';
 export const merge: IMerge = new Merge({ depth: 100 }).merge;
 export const clone = new Clone({ depth: 100 }).clone;
 
-export function isIterable (obj) {
+export function isIterable(obj: unknown): obj is Iterable<unknown> {
   if (obj === null || obj === undefined) {
     return false;
   }