+++ test/project/models/GetResponse.ts
@@ -1,0 +1,7 @@
+export type GetResponse = {
+  statusCode: number;
+  body: object;
+  headers: {
+    'Content-Type': string;
+  };
+};